requirements.txt
fastapi[standard]
python-multipart
PyMuPDF
sentence-transformers
faiss-cpu
numpy
transformers
streamlit
requests
pydantic
pandas